// API.ts
export type AggregatedStats = {
  __typename: "AggregatedStats",
  createdAt: string,
  id: string,
  losses: number,
  metadata?: string | null,
  owner?: string | null,
  statsKey: string,
  statsType: string,
  totalGames: number,
  updatedAt?: string | null,
  winRate: number,
  wins: number,
};

export type Deck = {
  __typename: "Deck",
  className: string,
  createdAt?: string | null,
  deckName: string,
  id: string,
  owner?: string | null,
  updatedAt: string,
  userId: string,
};

export type MatchRecord = {
  __typename: "MatchRecord",
  createdAt: string,
  id: string,
  isFirstPlayer: boolean,
  isWin: boolean,
  myDeckId: string,
  opponentClass: string,
  opponentDeckType: string,
  owner?: string | null,
  recordedAt: string,
  updatedAt: string,
  userId: string,
};

export type User = {
  __typename: "User",
  createdAt?: string | null,
  email: string,
  id: string,
  owner?: string | null,
  updatedAt: string,
};
// ...
export type ModelStringInput = {
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  beginsWith?: string | null,
  between?: Array< string | null > | null,
  contains?: string | null,
  eq?: string | null,
  ge?: string | null,
  gt?: string | null,
  le?: string | null,
  lt?: string | null,
  ne?: string | null,
  notContains?: string | null,
  size?: ModelSizeInput | null,
};
// ...
export type ModelIDInput = {
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  beginsWith?: string | null,
  between?: Array< string | null > | null,
  contains?: string | null,
  eq?: string | null,
  ge?: string | null,
  gt?: string | null,
  le?: string | null,
  lt?: string | null,
  ne?: string | null,
  notContains?: string | null,
  size?: ModelSizeInput | null,
};

export type ModelIntInput = {
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  between?: Array< number | null > | null,
  eq?: number | null,
  ge?: number | null,
  gt?: number | null,
  le?: number | null,
  lt?: number | null,
  ne?: number | null,
};

export type ModelFloatInput = {
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  between?: Array< number | null > | null,
  eq?: number | null,
  ge?: number | null,
  gt?: number | null,
  le?: number | null,
  lt?: number | null,
  ne?: number | null,
};
// ...
export type ModelBooleanInput = {
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  eq?: boolean | null,
  ne?: boolean | null,
};
// ...
export type ModelAggregatedStatsConditionInput = {
  and?: Array< ModelAggregatedStatsConditionInput | null > | null,
  createdAt?: ModelStringInput | null,
  losses?: ModelIntInput | null,
  metadata?: ModelStringInput | null,
  not?: ModelAggregatedStatsConditionInput | null,
  or?: Array< ModelAggregatedStatsConditionInput | null > | null,
  owner?: ModelStringInput | null,
  statsKey?: ModelStringInput | null,
  statsType?: ModelStringInput | null,
  totalGames?: ModelIntInput | null,
  updatedAt?: ModelStringInput | null,
  winRate?: ModelFloatInput | null,
  wins?: ModelIntInput | null,
};

export type CreateAggregatedStatsInput = {
  id?: string | null,
  losses: number,
  metadata?: string | null,
  statsKey: string,
  statsType: string,
  totalGames: number,
  updatedAt?: string | null,
  winRate: number,
  wins: number,
};

export type ModelDeckConditionInput = {
  and?: Array< ModelDeckConditionInput | null > | null,
  className?: ModelStringInput | null,
  createdAt?: ModelStringInput | null,
  deckName?: ModelStringInput | null,
  not?: ModelDeckConditionInput | null,
  or?: Array< ModelDeckConditionInput | null > | null,
  owner?: ModelStringInput | null,
  updatedAt?: ModelStringInput | null,
  userId?: ModelIDInput | null,
};

export type CreateDeckInput = {
  className: string,
  createdAt?: string | null,
  deckName: string,
  id?: string | null,
  userId: string,
};

export type ModelMatchRecordConditionInput = {
  and?: Array< ModelMatchRecordConditionInput | null > | null,
  createdAt?: ModelStringInput | null,
  isFirstPlayer?: ModelBooleanInput | null,
  isWin?: ModelBooleanInput | null,
  myDeckId?: ModelIDInput | null,
  not?: ModelMatchRecordConditionInput | null,
  opponentClass?: ModelStringInput | null,
  opponentDeckType?: ModelStringInput | null,
  or?: Array< ModelMatchRecordConditionInput | null > | null,
  owner?: ModelStringInput | null,
  recordedAt?: ModelStringInput | null,
  updatedAt?: ModelStringInput | null,
  userId?: ModelIDInput | null,
};

export type CreateMatchRecordInput = {
  id?: string | null,
  isFirstPlayer: boolean,
  isWin: boolean,
  myDeckId: string,
  opponentClass: string,
  opponentDeckType: string,
  recordedAt: string,
  userId: string,
};

export type ModelUserConditionInput = {
  and?: Array< ModelUserConditionInput | null > | null,
  createdAt?: ModelStringInput | null,
  email?: ModelStringInput | null,
  not?: ModelUserConditionInput | null,
  or?: Array< ModelUserConditionInput | null > | null,
  owner?: ModelStringInput | null,
  updatedAt?: ModelStringInput | null,
};

export type CreateUserInput = {
  createdAt?: string | null,
  email: string,
  id?: string | null,
};

export type DeleteAggregatedStatsInput = {
  id: string,
};

export type DeleteDeckInput = {
  id: string,
};

export type DeleteMatchRecordInput = {
  id: string,
};

export type DeleteUserInput = {
  id: string,
};

export type UpdateAggregatedStatsInput = {
  id: string,
  losses?: number | null,
  metadata?: string | null,
  statsKey?: string | null,
  statsType?: string | null,
  totalGames?: number | null,
  updatedAt?: string | null,
  winRate?: number | null,
  wins?: number | null,
};

export type UpdateDeckInput = {
  className?: string | null,
  createdAt?: string | null,
  deckName?: string | null,
  id: string,
  userId?: string | null,
};

export type UpdateMatchRecordInput = {
  id: string,
  isFirstPlayer?: boolean | null,
  isWin?: boolean | null,
  myDeckId?: string | null,
  opponentClass?: string | null,
  opponentDeckType?: string | null,
  recordedAt?: string | null,
  userId?: string | null,
};

export type UpdateUserInput = {
  createdAt?: string | null,
  email?: string | null,
  id: string,
};
// ...
export type CreateAggregatedStatsMutationVariables = {
  condition?: ModelAggregatedStatsConditionInput | null,
  input: CreateAggregatedStatsInput,
};

export type CreateAggregatedStatsMutation = {
  createAggregatedStats?:  {
    __typename: "AggregatedStats",
    createdAt: string,
    id: string,
    losses: number,
    metadata?: string | null,
    owner?: string | null,
    statsKey: string,
    statsType: string,
    totalGames: number,
    updatedAt?: string | null,
    winRate: number,
    wins: number,
  } | null,
};

export type CreateDeckMutationVariables = {
  condition?: ModelDeckConditionInput | null,
  input: CreateDeckInput,
};

export type CreateDeckMutation = {
  createDeck?:  {
    __typename: "Deck",
    className: string,
    createdAt?: string | null,
    deckName: string,
    id: string,
    owner?: string | null,
    updatedAt: string,
    userId: string,
  } | null,
};

export type CreateMatchRecordMutationVariables = {
  condition?: ModelMatchRecordConditionInput | null,
  input: CreateMatchRecordInput,
};

export type CreateMatchRecordMutation = {
  createMatchRecord?:  {
    __typename: "MatchRecord",
    createdAt: string,
    id: string,
    isFirstPlayer: boolean,
    isWin: boolean,
    myDeckId: string,
    opponentClass: string,
    opponentDeckType: string,
    owner?: string | null,
    recordedAt: string,
    updatedAt: string,
    userId: string,
  } | null,
};

export type CreateUserMutationVariables = {
  condition?: ModelUserConditionInput | null,
  input: CreateUserInput,
};

export type CreateUserMutation = {
  createUser?:  {
    __typename: "User",
    createdAt?: string | null,
    email: string,
    id: string,
    owner?: string | null,
    updatedAt: string,
  } | null,
};

export type DeleteAggregatedStatsMutationVariables = {
  condition?: ModelAggregatedStatsConditionInput | null,
  input: DeleteAggregatedStatsInput,
};

export type DeleteAggregatedStatsMutation = {
  deleteAggregatedStats?:  {
    __typename: "AggregatedStats",
    createdAt: string,
    id: string,
    losses: number,
    metadata?: string | null,
    owner?: string | null,
    statsKey: string,
    statsType: string,
    totalGames: number,
    updatedAt?: string | null,
    winRate: number,
    wins: number,
  } | null,
};

export type DeleteDeckMutationVariables = {
  condition?: ModelDeckConditionInput | null,
  input: DeleteDeckInput,
};

export type DeleteDeckMutation = {
  deleteDeck?:  {
    __typename: "Deck",
    className: string,
    createdAt?: string | null,
    deckName: string,
    id: string,
    owner?: string | null,
    updatedAt: string,
    userId: string,
  } | null,
};

export type DeleteMatchRecordMutationVariables = {
  condition?: ModelMatchRecordConditionInput | null,
  input: DeleteMatchRecordInput,
};

export type DeleteMatchRecordMutation = {
  deleteMatchRecord?:  {
    __typename: "MatchRecord",
    createdAt: string,
    id: string,
    isFirstPlayer: boolean,
    isWin: boolean,
    myDeckId: string,
    opponentClass: string,
    opponentDeckType: string,
    owner?: string | null,
    recordedAt: string,
    updatedAt: string,
    userId: string,
  } | null,
};

export type DeleteUserMutationVariables = {
  condition?: ModelUserConditionInput | null,
  input: DeleteUserInput,
};

export type DeleteUserMutation = {
  deleteUser?:  {
    __typename: "User",
    createdAt?: string | null,
    email: string,
    id: string,
    owner?: string | null,
    updatedAt: string,
  } | null,
};

export type UpdateAggregatedStatsMutationVariables = {
  condition?: ModelAggregatedStatsConditionInput | null,
  input: UpdateAggregatedStatsInput,
};

export type UpdateAggregatedStatsMutation = {
  updateAggregatedStats?:  {
    __typename: "AggregatedStats",
    createdAt: string,
    id: string,
    losses: number,
    metadata?: string | null,
    owner?: string | null,
    statsKey: string,
    statsType: string,
    totalGames: number,
    updatedAt?: string | null,
    winRate: number,
    wins: number,
  } | null,
};

export type UpdateDeckMutationVariables = {
  condition?: ModelDeckConditionInput | null,
  input: UpdateDeckInput,
};

export type UpdateDeckMutation = {
  updateDeck?:  {
    __typename: "Deck",
    className: string,
    createdAt?: string | null,
    deckName: string,
    id: string,
    owner?: string | null,
    updatedAt: string,
    userId: string,
  } | null,
};

export type UpdateMatchRecordMutationVariables = {
  condition?: ModelMatchRecordConditionInput | null,
  input: UpdateMatchRecordInput,
};

export type UpdateMatchRecordMutation = {
  updateMatchRecord?:  {
    __typename: "MatchRecord",
    createdAt: string,
    id: string,
    isFirstPlayer: boolean,
    isWin: boolean,
    myDeckId: string,
    opponentClass: string,
    opponentDeckType: string,
    owner?: string | null,
    recordedAt: string,
    updatedAt: string,
    userId: string,
  } | null,
};

export type UpdateUserMutationVariables = {
  condition?: ModelUserConditionInput | null,
  input: UpdateUserInput,
};

export type UpdateUserMutation = {
  updateUser?:  {
    __typename: "User",
    createdAt?: string | null,
    email: string,
    id: string,
    owner?: string | null,
    updatedAt: string,
  } | null,
};